+++ core-ng/src/main/java/core/framework/impl/web/HTTPServer.java
@@ -39,7 +39,8 @@ public class HTTPServer {
             if (httpPort != null) builder.addHttpListener(httpPort, "0.0.0.0");
             if (httpsPort != null) builder.addHttpsListener(httpsPort, "0.0.0.0", new SSLContextBuilder().build());
             builder.setHandler(new HTTPServerIOHandler(handler))
-                   .setServerOption(UndertowOptions.DECODE_URL, false);
+                   .setServerOption(UndertowOptions.DECODE_URL, false)
+                   .setServerOption(UndertowOptions.ENABLE_RFC6265_COOKIE_VALIDATION, true);
             server = builder.build();
             server.start();
         } finally {