+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 4.16.3 (9/5/2017 - )
 * mongo: tweak mongo encoder and decoder code gen   
 * http: update undertow to 1.4.20, tweak the cookies config
+* search: update es to 5.6 
 
 ### 4.16.2 (9/1/2017 - 9/4/2017)
 * bug: fix oracle pagination query param value